EPnP pose solver Lepetit et al., Change to Point2f, Point3f for plugin integration

+++ smll/FaceLib/FaceLandmarks.h
@@ -83,12 +83,12 @@ namespace FaceLib {
 		*/
 		void DetectLandmarks(cv::Mat& image, dlib::rectangle& face, std::vector<dlib::point>& landmarks);
 
-		void DetectPose(std::vector<cv::Point2d>& landmarks2D, std::vector<cv::Point3d>& landmarks3D, cv::Mat& K, cv::Mat& D, cv::Mat& R, cv::Mat& t);
+		void DetectPose(std::vector<cv::Point2f>& landmarks2D, std::vector<cv::Point3f>& landmarks3D, cv::Mat& K, cv::Mat& D, cv::Mat& R, cv::Mat& t);
 		void DetectPose(std::vector<cv::Point2d>& landmarks2D, cv::Mat& K, cv::Mat& D, cv::Mat& R, cv::Mat& t);
 
 	private:
 		dlib::shape_predictor _landmarksPredictor;
-		std::vector<cv::Point3f> _landmarks3D;
+		std::vector<cv::Point3d> _landmarks3D;
 
 		void InitLandmarks3D();
 		void CorrectPoseFlip(cv::Mat& R, cv::Mat& t);
